iamodel/service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
openai
python-dotenv
